Remove hard-coded path to test executable.
Such paths are brittle, due to npm deduping, mocha is not guaranteed to
be at this location, it may be higher in the tree.

node_modules/.bin is added to the PATH env var for npm scripts so
"mocha" on its own is both shorter and will more robustly be able to
locate the mocha executable.