Fixing node header issue.
Node adds a header of 62 characters when code is required through a
native module. If we are in node, the line number is 1 and the next file
is the native 'module.js', we subtract that header. I verified the
solution for https://github.com/evanw/node-source-map-support/issues/36
by reproducing the problem. The stack trace is now:

```
user:node-source-map-support user$ node error.js
1

/Users/user/Documents/node-source-map-support/uglifysanity.js:2
goober();
^
ReferenceError: goober is not defined
at Object.<anonymous>
(/Users/user/Documents/node-source-map-support/uglifysanity.js:2:1)
```