fix: remove --debug flag for browserify.
According to the spec (https://goo.gl/UQJKOW) the source map comment should be
at the very end of the file. Browser's adhere to that, i.e. use the last source
map in the file.

Browserify's --debug adds a source map at the bottom of the compiled file, but
currently fails to combine that with the source map referenced in script.js.
This is a bug in Browserify, see issue
https://github.com/substack/node-browserify/issues/772.

Our test case thus only worked by accident, for the user in the browser only
"script.js" shows up, not "script.coffee" as it should.

So for the time being, we should just remove the --debug flag from the build and
re-enable once browserify fixes that issue.