Fix RangeError on large files
When the source file and/or the source map URL is sufficiently
large (think several MB) the regexp search to find the URL can run into
this issue: https://code.google.com/p/v8/issues/detail?id=3878

Change the regexp to look just for the prefix and extract the URL using
substring, which scales better.

This now doesn't pay attention to single and double quotes anymore,
reading the spec there seems to be no need for it.