Allow for multiple install calls
This allows for multiple libraries to utilize this without stepping on each other’s toes.

This makes the retrieve helpers additive, giving priority to the last retrieve method registered. In the event of a miss, the next method is called up to and including the default behavior. The `overrideRetriveFile` and `overrideRetrieveSourceMap` flags are added if this behavior is not desired but this is something that most libraries should avoid to be good citizens.

Flagged behaviors will now be on if any of the install calls dictate that they would be.

Fix for #91