Fix Uncaught Type Error
Fixes https://github.com/evanw/node-source-map-support/issues/167

Since version 0.4.9, we require `fs` but it should fail if we are in
a browser. However, `browser-source-map-support.js` is bundled with
browserify which converts unsupported builtins to empty object. See
https://github.com/substack/node-browserify/blob/8f5ebbf505/lib/builtins.js#L12
This PR checks if required `fs` contains `readFileSync` which is part
of `fs` API.