Be resilient against readFileSync failures
Generating a call stack when using this library can trigger a call to
fs.readFileSync.  That call can sometimes fail, resulting in a very
confusing stack trace (or none at all, which is even more confusing).

This patch wraps calls to fs.readFileSync() in a try/catch.  If the read
fails, then the file contents are omitted, but this is far better than
total failure.