fix: replace build.js curl call with node https
This was occasionally causing my machine to spin on CPU indefinitely
while trying to download the Closure Compiler output.

It would perhaps be better to investigate what is wrong with curl on my
machine, but since Node has a pretty nice built-in HTTPS client, I
figured it'd be easier to just cut out the external system dependency
entirely.